Create new_wordle.java
Updated framework with implemented functions

TO-DO: see lines 83-90
In the loop of 6 words from the user, we need to check if the current guess is the correct guess and exit the loop if it is. We also need to conclude the game with a win/loss message and any other cosmetic details we want to add.